requirements_fixed.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
bcrypt==4.0.1
passlib==1.7.4
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
email-validator==2.1.0
itsdangerous==2.1.2
aiosmtplib==2.0.1
pandas==2.1.3
openpyxl==3.1.2
